requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
Brotli==1.0.9
certifi==2022.12.7
chardet==5.2.0
charset-normalizer==3.0.1
click==8.1.3
ConfigArgParse==1.5.3
fastapi==0.104.1
ffmpeg==1.4
Flask==2.2.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
frozenlist==1.4.0
gevent==22.10.2
geventhttpclient==2.0.8
greenlet==2.0.2
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
locust==2.15.1
MarkupSafe==2.1.2
msgpack==1.0.4
multidict==6.0.4
polling==0.3.2
psutil==5.9.4
pydantic==2.5.2
pydantic_core==2.14.5
python-gitlab==3.13.0
pyzmq==25.0.0
requests==2.28.2
requests-toolbelt==0.10.1
roundrobin==0.0.4
routers==0.10.1
sentry-sdk==1.39.0
six==1.16.0
slack-bolt==1.18.0
slack-sdk==3.23.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.9.0
urllib3==1.26.14
uvicorn==0.24.0.post1
Werkzeug==2.2.3
yarl==1.9.2
zope.event==4.6
zope.interface==5.5.2
